+++ src/ast.rs
@@ -1394,18 +1394,9 @@ pub struct CustomExpr {
     /// List of keywords.
     pub keywords: StaticVec<Expr>,
     /// Is the current [`Scope`][crate::Scope] modified?
-    pub scope_changed: bool,
+    pub scope_may_be_changed: bool,
     /// List of tokens actually parsed.
     pub tokens: StaticVec<Identifier>,
-}
-
-impl CustomExpr {
-    /// Convert this into a [`Expr::Custom`].
-    #[inline(always)]
-    #[must_use]
-    pub fn into_custom_syntax_expr(self, pos: Position) -> Expr {
-        Expr::Custom(self.into(), pos)
-    }
 }
 
 /// _(INTERNALS)_ A binary expression.