+++ src/custom_syntax.rs
@@ -156,33 +156,34 @@ pub struct CustomSyntax {
     /// Custom syntax implementation function.
     pub func: Shared<FnCustomSyntaxEval>,
     /// Any variables added/removed in the scope?
-    pub scope_changed: bool,
+    pub scope_may_be_changed: bool,
 }
 
 impl Engine {
     /// Register a custom syntax with the [`Engine`].
     ///
     /// * `keywords` holds a slice of strings that define the custom syntax.  
-    /// * `scope_changed` specifies variables have been added/removed by this custom syntax.
+    /// * `scope_may_be_changed` specifies variables _may_ be added/removed by this custom syntax.
     /// * `func` is the implementation function.
     ///
-    /// # Caveat - Do not change beyond block scope
+    /// ## Note on `scope_may_be_changed`
     ///
-    /// If `scope_changed` is `true`, then the current [`Scope`][crate::Scope] is assumed to be
-    /// modified by this custom syntax.
+    /// If `scope_may_be_changed` is `true`, then _size_ of the current [`Scope`][crate::Scope]
+    /// _may_ be modified by this custom syntax.
     ///
-    /// Adding new variables and/or removing variables are allowed. Simply modifying the values of
-    /// variables does NOT count, so `false` should be passed in this case.
+    /// Adding new variables and/or removing variables count.
     ///
-    /// However, only variables declared within the current _block scope_ should be touched,
-    /// since they all go away at the end of the block.
+    /// Simply modifying the values of existing variables does NOT count, as the _size_ of the
+    /// current [`Scope`][crate::Scope] is unchanged, so `false` should be passed.
     ///
-    /// Variables in parent blocks should be left untouched as they persist beyond the current block.
+    /// Replacing one variable with another (i.e. adding a new variable and removing one variable at
+    /// the same time so that the total _size_ of the [`Scope`][crate::Scope] is unchanged) also
+    /// does NOT count, so `false` should be passed.
     #[must_use]
     pub fn register_custom_syntax<S: AsRef<str> + Into<Identifier>>(
         &mut self,
         keywords: &[S],
-        scope_changed: bool,
+        scope_may_be_changed: bool,
         func: impl Fn(&mut EvalContext, &[Expression]) -> RhaiResult + SendSync + 'static,
     ) -> Result<&mut Self, ParseError> {
         let keywords = keywords.as_ref();
@@ -289,7 +290,7 @@ impl Engine {
                     Ok(Some(segments[stream.len()].clone()))
                 }
             },
-            scope_changed,
+            scope_may_be_changed,
             func,
         );
 
@@ -301,7 +302,7 @@ impl Engine {
     ///
     /// This function is very low level.
     ///
-    /// * `scope_changed` specifies variables have been added/removed by this custom syntax.
+    /// * `scope_may_be_changed` specifies variables have been added/removed by this custom syntax.
     /// * `parse` is the parsing function.
     /// * `func` is the implementation function.
     ///
@@ -313,7 +314,7 @@ impl Engine {
         parse: impl Fn(&[ImmutableString], &str) -> Result<Option<ImmutableString>, ParseError>
             + SendSync
             + 'static,
-        scope_changed: bool,
+        scope_may_be_changed: bool,
         func: impl Fn(&mut EvalContext, &[Expression]) -> RhaiResult + SendSync + 'static,
     ) -> &mut Self {
         self.custom_syntax.insert(
@@ -321,7 +322,7 @@ impl Engine {
             CustomSyntax {
                 parse: Box::new(parse),
                 func: (Box::new(func) as Box<FnCustomSyntaxEval>).into(),
-                scope_changed,
+                scope_may_be_changed,
             }
             .into(),
         );