+++ src/engine.rs
@@ -24,6 +24,7 @@ use std::{
     collections::{BTreeMap, BTreeSet},
     fmt,
     hash::{Hash, Hasher},
+    iter::{FromIterator, Rev, Zip},
     num::{NonZeroU8, NonZeroUsize},
     ops::{Deref, DerefMut},
 };
@@ -62,6 +63,12 @@ pub struct Imports {
 }
 
 impl Imports {
+    /// Create a new stack of imported [modules][Module].
+    #[inline(always)]
+    #[must_use]
+    pub fn new() -> Self {
+        Default::default()
+    }
     /// Get the length of this stack of imported [modules][Module].
     #[inline(always)]
     #[must_use]
@@ -74,20 +81,20 @@ impl Imports {
     pub fn is_empty(&self) -> bool {
         self.keys.is_empty()
     }
-    /// Get the imported [modules][Module] at a particular index.
+    /// Get the imported [module][Module] at a particular index.
     #[inline(always)]
     #[must_use]
     pub fn get(&self, index: usize) -> Option<Shared<Module>> {
         self.modules.get(index).cloned()
     }
-    /// Get the imported [modules][Module] at a particular index.
+    /// Get the imported [module][Module] at a particular index.
     #[allow(dead_code)]
     #[inline(always)]
     #[must_use]
     pub(crate) fn get_mut(&mut self, index: usize) -> Option<&mut Shared<Module>> {
         self.modules.get_mut(index)
     }
-    /// Get the index of an imported [modules][Module] by name.
+    /// Get the index of an imported [module][Module] by name.
     #[inline(always)]
     #[must_use]
     pub fn find(&self, name: &str) -> Option<usize> {
@@ -97,7 +104,7 @@ impl Imports {
             .rev()
             .find_map(|(i, key)| if key == name { Some(i) } else { None })
     }
-    /// Push an imported [modules][Module] onto the stack.
+    /// Push an imported [module][Module] onto the stack.
     #[inline(always)]
     pub fn push(&mut self, name: impl Into<Identifier>, module: impl Into<Shared<Module>>) {
         self.keys.push(name.into());
@@ -134,15 +141,6 @@ impl Imports {
     pub(crate) fn scan_raw(&self) -> impl Iterator<Item = (&Identifier, &Shared<Module>)> {
         self.keys.iter().zip(self.modules.iter())
     }
-    /// Get a consuming iterator to this stack of imported [modules][Module] in reverse order.
-    #[inline(always)]
-    #[must_use]
-    pub fn into_iter(self) -> impl Iterator<Item = (Identifier, Shared<Module>)> {
-        self.keys
-            .into_iter()
-            .rev()
-            .zip(self.modules.into_iter().rev())
-    }
     /// Does the specified function hash key exist in this stack of imported [modules][Module]?
     #[allow(dead_code)]
     #[inline(always)]
@@ -150,7 +148,7 @@ impl Imports {
     pub fn contains_fn(&self, hash: u64) -> bool {
         self.modules.iter().any(|m| m.contains_qualified_fn(hash))
     }
-    /// Get specified function via its hash key.
+    /// Get the specified function via its hash key from this stack of imported [modules][Module].
     #[inline(always)]
     #[must_use]
     pub fn get_fn(&self, hash: u64) -> Option<(&CallableFunction, Option<&Identifier>)> {
@@ -167,7 +165,8 @@ impl Imports {
     pub fn contains_iter(&self, id: TypeId) -> bool {
         self.modules.iter().any(|m| m.contains_qualified_iter(id))
     }
-    /// Get the specified [`TypeId`][std::any::TypeId] iterator.
+    /// Get the specified [`TypeId`][std::any::TypeId] iterator from this stack of imported
+    /// [modules][Module].
     #[inline(always)]
     #[must_use]
     pub fn get_iter(&self, id: TypeId) -> Option<IteratorFn> {
@@ -175,6 +174,37 @@ impl Imports {
             .iter()
             .rev()
             .find_map(|m| m.get_qualified_iter(id))
+    }
+}
+
+impl IntoIterator for Imports {
+    type Item = (Identifier, Shared<Module>);
+    type IntoIter =
+        Zip<Rev<smallvec::IntoIter<[Identifier; 4]>>, Rev<smallvec::IntoIter<[Shared<Module>; 4]>>>;
+
+    #[inline(always)]
+    fn into_iter(self) -> Self::IntoIter {
+        self.keys
+            .into_iter()
+            .rev()
+            .zip(self.modules.into_iter().rev())
+    }
+}
+
+impl<K: Into<Identifier>, M: Into<Shared<Module>>> FromIterator<(K, M)> for Imports {
+    fn from_iter<T: IntoIterator<Item = (K, M)>>(iter: T) -> Self {
+        let mut lib = Self::new();
+        lib.extend(iter);
+        lib
+    }
+}
+
+impl<K: Into<Identifier>, M: Into<Shared<Module>>> Extend<(K, M)> for Imports {
+    fn extend<T: IntoIterator<Item = (K, M)>>(&mut self, iter: T) {
+        iter.into_iter().for_each(|(k, m)| {
+            self.keys.push(k.into());
+            self.modules.push(m.into());
+        })
     }
 }
 
@@ -348,7 +378,7 @@ impl<'a> Target<'a> {
     #[allow(dead_code)]
     #[inline(always)]
     #[must_use]
-    pub fn is_ref(&self) -> bool {
+    pub const fn is_ref(&self) -> bool {
         match self {
             Self::RefMut(_) => true,
             #[cfg(not(feature = "no_closure"))]
@@ -363,7 +393,7 @@ impl<'a> Target<'a> {
     /// Is the `Target` a temp value?
     #[inline(always)]
     #[must_use]
-    pub fn is_temp_value(&self) -> bool {
+    pub const fn is_temp_value(&self) -> bool {
         match self {
             Self::RefMut(_) => false,
             #[cfg(not(feature = "no_closure"))]
@@ -578,7 +608,12 @@ impl<T: Into<Dynamic>> From<T> for Target<'_> {
     }
 }
 
-/// An entry in a function resolution cache.
+/// _(INTERNALS)_ An entry in a function resolution cache.
+/// Exported under the `internals` feature only.
+///
+/// # Volatile Data Structure
+///
+/// This type is volatile and may change.
 #[derive(Debug, Clone)]
 pub struct FnResolutionCacheEntry {
     /// Function.
@@ -587,7 +622,12 @@ pub struct FnResolutionCacheEntry {
     pub source: Option<Identifier>,
 }
 
-/// A function resolution cache.
+/// _(INTERNALS)_ A function resolution cache.
+/// Exported under the `internals` feature only.
+///
+/// # Volatile Data Structure
+///
+/// This type is volatile and may change.
 pub type FnResolutionCache = BTreeMap<u64, Option<Box<FnResolutionCacheEntry>>>;
 
 /// _(INTERNALS)_ A type that holds all the current states of the [`Engine`].
@@ -597,15 +637,15 @@ pub type FnResolutionCache = BTreeMap<u64, Option<Box<FnResolutionCacheEntry>>>;
 ///
 /// This type is volatile and may change.
 #[derive(Debug, Clone, Default)]
-pub struct State {
+pub struct EvalState {
     /// Source of the current context.
     pub source: Option<Identifier>,
-    /// Normally, access to variables are parsed with a relative offset into the scope to avoid a lookup.
-    /// In some situation, e.g. after running an `eval` statement, subsequent offsets become mis-aligned.
-    /// When that happens, this flag is turned on to force a scope lookup by name.
-    pub always_search: bool,
-    /// Level of the current scope.  The global (root) level is zero, a new block
-    /// (or function call) is one level higher, and so on.
+    /// Normally, access to variables are parsed with a relative offset into the [`Scope`] to avoid a lookup.
+    /// In some situation, e.g. after running an `eval` statement, subsequent offsets may become mis-aligned.
+    /// When that happens, this flag is turned on to force a [`Scope`] search by name.
+    pub always_search_scope: bool,
+    /// Level of the current scope.  The global (root) level is zero, a new block (or function call)
+    /// is one level higher, and so on.
     pub scope_level: usize,
     /// Number of operations performed.
     pub operations: u64,
@@ -614,15 +654,30 @@ pub struct State {
     /// Embedded module resolver.
     #[cfg(not(feature = "no_module"))]
     pub resolver: Option<Shared<crate::module::resolvers::StaticModuleResolver>>,
-    /// Function resolution cache and free list.
-    fn_resolution_caches: StaticVec<FnResolutionCache>,
+    /// Stack of function resolution caches.
+    fn_resolution_caches: Vec<FnResolutionCache>,
 }
 
-impl State {
+impl EvalState {
+    /// Create a new [`EvalState`].
+    #[inline(always)]
+    #[must_use]
+    pub const fn new() -> Self {
+        Self {
+            source: None,
+            always_search_scope: false,
+            scope_level: 0,
+            operations: 0,
+            modules: 0,
+            #[cfg(not(feature = "no_module"))]
+            resolver: None,
+            fn_resolution_caches: Vec::new(),
+        }
+    }
     /// Is the state currently at global (root) level?
     #[inline(always)]
     #[must_use]
-    pub fn is_global(&self) -> bool {
+    pub const fn is_global(&self) -> bool {
         self.scope_level == 0
     }
     /// Get a mutable reference to the current function resolution cache.
@@ -729,13 +784,13 @@ pub struct EvalContext<'a, 'x, 'px, 'm, 's, 'b, 't, 'pt> {
     pub(crate) engine: &'a Engine,
     pub(crate) scope: &'x mut Scope<'px>,
     pub(crate) mods: &'m mut Imports,
-    pub(crate) state: &'s mut State,
+    pub(crate) state: &'s mut EvalState,
     pub(crate) lib: &'b [&'b Module],
     pub(crate) this_ptr: &'t mut Option<&'pt mut Dynamic>,
     pub(crate) level: usize,
 }
 
-impl<'x, 'px> EvalContext<'_, 'x, 'px, '_, '_, '_, '_, '_> {
+impl<'x, 'px, 'pt> EvalContext<'_, 'x, 'px, '_, '_, '_, '_, 'pt> {
     /// The current [`Engine`].
     #[inline(always)]
     #[must_use]
@@ -795,6 +850,12 @@ impl<'x, 'px> EvalContext<'_, 'x, 'px, '_, '_, '_, '_, '_> {
     #[must_use]
     pub fn this_ptr(&self) -> Option<&Dynamic> {
         self.this_ptr.as_ref().map(|v| &**v)
+    }
+    /// Mutable reference to the current bound `this` pointer, if any.
+    #[inline(always)]
+    #[must_use]
+    pub fn this_ptr_mut(&mut self) -> Option<&mut &'pt mut Dynamic> {
+        self.this_ptr.as_mut()
     }
     /// The current nesting level of function calls.
     #[inline(always)]
@@ -1002,13 +1063,13 @@ impl Engine {
     pub(crate) fn search_imports(
         &self,
         mods: &Imports,
-        state: &mut State,
+        state: &mut EvalState,
         namespace: &NamespaceRef,
     ) -> Option<Shared<Module>> {
         let root = &namespace[0].name;
 
         // Qualified - check if the root module is directly indexed
-        let index = if state.always_search {
+        let index = if state.always_search_scope {
             None
         } else {
             namespace.index()
@@ -1037,7 +1098,7 @@ impl Engine {
         &self,
         scope: &'s mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &'s mut Option<&mut Dynamic>,
         expr: &Expr,
@@ -1087,7 +1148,7 @@ impl Engine {
         &self,
         scope: &'s mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &'s mut Option<&mut Dynamic>,
         expr: &Expr,
@@ -1103,7 +1164,7 @@ impl Engine {
                     EvalAltResult::ErrorUnboundThis(*pos).into()
                 }
             }
-            _ if state.always_search => (0, expr.position()),
+            _ if state.always_search_scope => (0, expr.position()),
             Expr::Variable(Some(i), pos, _) => (i.get() as usize, *pos),
             Expr::Variable(None, pos, v) => (v.0.map(NonZeroUsize::get).unwrap_or(0), *pos),
             _ => unreachable!("Expr::Variable expected, but gets {:?}", expr),
@@ -1160,7 +1221,7 @@ impl Engine {
     fn eval_dot_index_chain_helper(
         &self,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         target: &mut Target,
@@ -1592,7 +1653,7 @@ impl Engine {
         &self,
         scope: &mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         expr: &Expr,
@@ -1656,7 +1717,7 @@ impl Engine {
         &self,
         scope: &mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         expr: &Expr,
@@ -1785,7 +1846,7 @@ impl Engine {
     fn get_indexed_mut<'t>(
         &self,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         target: &'t mut Dynamic,
         mut idx: Dynamic,
@@ -1843,7 +1904,7 @@ impl Engine {
             #[cfg(not(feature = "no_object"))]
             Dynamic(Union::Map(map, _, _)) => {
                 // val_map[idx]
-                let index = &*idx.read_lock::<ImmutableString>().ok_or_else(|| {
+                let index = idx.read_lock::<ImmutableString>().ok_or_else(|| {
                     self.make_type_mismatch_err::<ImmutableString>(idx.type_name(), idx_pos)
                 })?;
 
@@ -1957,7 +2018,7 @@ impl Engine {
         &self,
         scope: &mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         expr: &Expr,
@@ -2158,7 +2219,7 @@ impl Engine {
         &self,
         scope: &mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         statements: &[Stmt],
@@ -2170,7 +2231,7 @@ impl Engine {
         }
 
         let mut _extra_fn_resolution_cache = false;
-        let prev_always_search = state.always_search;
+        let prev_always_search_scope = state.always_search_scope;
         let prev_scope_len = scope.len();
         let prev_mods_len = mods.len();
 
@@ -2223,7 +2284,7 @@ impl Engine {
 
             // The impact of new local variables goes away at the end of a block
             // because any new variables introduced will go out of scope
-            state.always_search = prev_always_search;
+            state.always_search_scope = prev_always_search_scope;
         }
 
         result
@@ -2235,7 +2296,7 @@ impl Engine {
     pub(crate) fn eval_op_assignment(
         &self,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         op_info: Option<OpAssignment>,
         op_pos: Position,
@@ -2310,7 +2371,7 @@ impl Engine {
         &self,
         scope: &mut Scope,
         mods: &mut Imports,
-        state: &mut State,
+        state: &mut EvalState,
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         stmt: &Stmt,
@@ -3077,7 +3138,7 @@ impl Engine {
     #[must_use]
     pub(crate) fn inc_operations(
         &self,
-        state: &mut State,
+        state: &mut EvalState,
         pos: Position,
     ) -> Result<(), Box<EvalAltResult>> {
         state.operations += 1;